店情報参照 エンティティをFootPrintに修正
